# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quizerland API",
        "version": "1.0"
    },
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Вход",
                "description": "Вход в систему по логину и паролю",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Неверный логин и/или пароль"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Выход",
                "description": "Выход из системы",
                "operationId": "03afb12b342c2ea43bdc1de4f39e91a8",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Регистрация",
                "description": "Регистрация нового пользователя в системе",
                "operationId": "573de1fed352c1205a32c4d1b9877375",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jason Statham"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "fake@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secure-password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secure-password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Невалидные данные"
                    }
                }
            }
        },
        "/api/quizzes": {
            "get": {
                "tags": [
                    "Тесты"
                ],
                "summary": "Список тестов",
                "description": "Получить список тестов с пагинацией",
                "operationId": "eb9ba15b35ba53ab91ba8b1a4e215e29",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Хеш для получения следующего блока тестов",
                        "schema": {
                            "type": "string",
                            "example": "$2fssfhu9ewagh97e9aus90few09ayfds9fa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "next_cursor": {
                                            "type": "string",
                                            "example": "$2fssfhu9ewagh97e9aus90few09ayfds9fa",
                                            "nullable": true
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/QuizSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Тесты"
                ],
                "summary": "Создать тест",
                "description": "Создать тест",
                "operationId": "2513a2f792e5c3a5bccd78eb9b4e3786",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuizCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quizzes/:id": {
            "get": {
                "tags": [
                    "Тесты"
                ],
                "summary": "Тест",
                "description": "Получить тест по идентификатору",
                "operationId": "7c6b463a11099b05d6bfc1db7d5de19f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор теста",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuizSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Пользователь",
                "description": "Получить данные авторизованного пользователя",
                "operationId": "59d04a61de17524a2768d4d7750d2d19",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Jason Statham"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "fake@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-05-20T12:45:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-05-20T12:45:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "QuizCreateRequest": {
                "properties": {
                    "title": {
                        "description": "Название теста",
                        "type": "string",
                        "example": "Тест на знание анатомии"
                    },
                    "description": {
                        "description": "Описание теста",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit emmet"
                    },
                    "questions": {
                        "description": "Вопросы",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionForQuizCreateRequest"
                        }
                    }
                },
                "type": "object"
            },
            "QuestionForQuizCreateRequest": {
                "properties": {
                    "text": {
                        "description": "Текст вопроса",
                        "type": "string",
                        "example": "Какой орган человека похож на грецкий орех?"
                    },
                    "correct_answer_index": {
                        "description": "Индекс правильного ответа",
                        "type": "integer"
                    },
                    "answers": {
                        "description": "Ответы",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Мозг"
                        }
                    }
                },
                "type": "object"
            },
            "AnswerSchema": {
                "properties": {
                    "id": {
                        "description": "-------- Columns ---------",
                        "type": "integer"
                    },
                    "text": {
                        "type": "string",
                        "example": "Текст ответа..."
                    }
                },
                "type": "object"
            },
            "QuestionSchema": {
                "properties": {
                    "id": {
                        "description": "-------- Columns ---------",
                        "type": "integer"
                    },
                    "text": {
                        "type": "string",
                        "example": "Текст ответа..."
                    },
                    "correct_answer_id": {
                        "type": "number"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnswerSchema"
                        }
                    }
                },
                "type": "object"
            },
            "QuizSchema": {
                "properties": {
                    "id": {
                        "description": "-------- Columns ---------",
                        "type": "integer"
                    },
                    "title": {
                        "type": "string",
                        "example": "Тест по географии"
                    },
                    "description": {
                        "type": "string",
                        "example": "Геогра́фия — комплекс естественных и общественных наук"
                    },
                    "author_id": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-05-20T12:45:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-05-20T12:45:00.000000Z"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionSchema"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}